+++ engines/azure/__init__.py
@@ -1,0 +1,3 @@
+"""Azure Cognitive Services TTS 引擎"""
+
+from .engine import AzureEngine