+++ engines/volcengine/__init__.py
@@ -1,0 +1,3 @@
+"""火山引擎 TTS 引擎"""
+
+from .engine import VolcengineEngine